refactor generic functionality out of ssa-specific code
