+++ idfa_livevars.c
@@ -3,54 +3,6 @@
 #include "symtab_basicblock.h"
 #include "util.h"
 
-ssize_t compareTacOperandIgnoreSsaNumber(void *dataA, void *dataB)
-{
-    struct TACOperand *operandA = dataA;
-    struct TACOperand *operandB = dataB;
-
-    if (Type_Compare(&operandA->type, &operandB->type))
-    {
-        return 1;
-    }
-
-    if (Type_Compare(&operandA->castAsType, &operandB->castAsType))
-    {
-        return 1;
-    }
-
-    if ((operandA->permutation != vp_literal && (operandB->permutation != vp_literal)))
-    {
-        if (strcmp(operandA->name.str, operandB->name.str))
-        {
-            return 1;
-        }
-    }
-    else if (operandA->permutation != operandB->permutation)
-    {
-        return 1;
-    }
-
-    return 0;
-}
-
-ssize_t compareTacOperand(void *dataA, void *dataB)
-{
-    struct TACOperand *operandA = dataA;
-    struct TACOperand *operandB = dataB;
-
-    if (compareTacOperandIgnoreSsaNumber(dataA, dataB))
-    {
-        return 1;
-    }
-
-    if (operandA->ssaNumber != operandB->ssaNumber)
-    {
-        return 1;
-    }
-
-    return 0;
-}
-
 struct Set *liveVars_transfer(struct Idfa *idfa, struct BasicBlock *block, struct Set *facts)
 {
     struct Set *transferred = Set_Copy(idfa->facts.gen[block->labelNum]);
@@ -111,29 +63,13 @@ void liveVars_findGenKills(struct Idfa *idfa)
     }
 }
 
-void printTACOperand(void *operandData)
-{
-    struct TACOperand *operand = operandData;
-    char *typeName = Type_GetName(&operand->type);
-    printf("%s", typeName);
-    if (operand->castAsType.basicType != vt_null)
-
-    {
-        char *castAsTypeName = Type_GetName(&operand->castAsType);
-        printf("(%s)", castAsTypeName);
-        free(castAsTypeName);
-    }
-    printf(" %s_%zu", operand->name.str, operand->ssaNumber);
-    free(typeName);
-}
-
 struct Idfa *analyzeLiveVars(struct IdfaContext *context)
 {
     struct Idfa *liveVarsIdfa = Idfa_Create(context,
                                             liveVars_transfer,
                                             liveVars_findGenKills,
                                             d_forwards,
-                                            compareTacOperand,
+                                            TACOperand_Compare,
                                             printTACOperand,
                                             Set_Union);
